pyproject.toml
[project]
name = "google_workspace_sync"
version = "0.1.0"
description = "One-shot pull sync for Google Drive folder mirror and Sheets sync"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.13.1",
]

[project.scripts]
google-workspace-sync = "google_workspace_sync.cli:main"

[project.optional-dependencies]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.5",
    "ty>=0.0.1a27",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "models.py"
]


# Same as Black.
line-length = 88
target-version = "py313"


[tool.ruff.lint]
#https://beta.ruff.rs/docs/rules/
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "G",
    "SIM1",
    "PLE",
    "N802", "N803", "N815", "N816","N806",
    "ANN",
    "C901",
    "UP",  # pyupgrade,
    "PLR0915"
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "G004",
    "ANN003",
    "B904",
    "G201"
]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

[tool.ruff.lint.pylint]
max-statements = 60

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]
